Backend/requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary
dj-database-url
whitenoise
django-cors-headers